refactor: 네이밍 앞에 -Memo-` prefix로 통일
