+++ packages/web/src/app/[lng]/(auth)/memos/_components/index.ts
@@ -3,6 +3,6 @@ export { default as MemoDialog } from './MemoDialog';
 export { default as MemoSidebar } from './MemoSidebar';
 export { default as MemoSidebarTrigger } from './MemoSidebarTrigger';
 export { default as MemoView } from './MemoView';
-export { default as SearchFormProvider } from './SearchFormProvider';
-export { default as SearchForm } from './SearchForm';
+export { default as MemoSearchFormProvider } from './MemoSearchFormProvider';
+export { default as MemoSearchForm } from './MemoSearchForm';
 export { default as InitSentryUserInfo } from './InitSentryUserInfo';
